packaging: trim tarball 266→103 KiB (27→10 files) by moving maintainer-only sources — keccak/codec/ntt unit tests, kani_proofs.rs (formal-verification harnesses), adversarial_proptests.rs (property-test battery + Lean-spec refinement sample), spec_pins.rs (spec-constant pins), miri_unsafe_paths.rs — out of src/ into a single internal-tests/ directory, pulled into the lib via `#[cfg(test)] #[path = "../internal-tests/<name>.rs"] mod ...;` (Kani harnesses use `#[cfg(kani)]`). [package.exclude] now lists four directories (internal-tests, formal_verification, proptest-regressions, .github) instead of enumerating each file. No `verification` feature: `cargo test` from the git checkout compiles all 70 lib tests; `cargo test` against the published tarball doesn't compile (test files absent by design — tests are intended to run from the git checkout). CI and scripts/check_all.sh no longer pass --features verification. Split the previous monolithic `formal_verification.rs` into kani_proofs.rs (`#[cfg(kani)]`, the actual `#[kani::proof]` harnesses) and adversarial_proptests.rs (`#[cfg(test)]`, randomized property tests) so the module name matches the contents. Sibling rename in host-tests/tests/: `constants_match_spec.rs` → `pqclean_wire_format.rs`, to disambiguate from the in-lib spec_pins (in-lib: pins internal arithmetic constants against the math spec; host-tests: cross-checks public wire bytes against PQClean output). #[cfg(test)] alone doesn't shrink the tarball — cargo packages source bytes regardless of cfg state.
