+++ Cargo.toml
@@ -9,6 +9,21 @@ keywords = ["falcon", "post-quantum", "signature", "solana", "no-std"]
 categories = ["cryptography", "no-std"]
 readme = "README.md"
 
+# Keep the published tarball lean. Maintainer-only sources — Rust unit
+# tests, Kani harnesses, the spec-pin and Miri shims — live under
+# `internal-tests/` and are pulled into the lib via `#[cfg(test)] #[path
+# = "../internal-tests/<name>.rs"]` declarations. Excluding that
+# directory plus the Lean proofs, proptest regressions, and CI config
+# trims the tarball to just the production sources. `cargo test` on the
+# *published* tarball won't compile (test files are absent by design) —
+# tests are intended to run from the git checkout.
+exclude = [
+    "internal-tests",
+    "formal_verification",
+    "proptest-regressions",
+    ".github",
+]
+
 [lib]
 crate-type = ["rlib"]
